+++ inst/doc/translate.R
@@ -30,7 +30,7 @@ cli_code <-
 translated <- cms_translate(cli_code)
 summary(translated)
 
-## ----download-translation, eval=eval_chunks, message=FALSE, fig.width=4, fig.height=4, fig.alt="Data downloaded using translated query code"----
+## ----download-translation, eval=eval_chunks, echo=eval_chunks, message=FALSE, fig.width=4, fig.height=4, fig.alt="Data downloaded using translated query code"----
 result <- do.call(cms_download_subset, translated)
 plot(result, col = hcl.colors(100), axes = TRUE)